server/src/main/java/fr/istic/gm/weplan/server/controller/RegionController.java
package fr.istic.gm.weplan.server.controller;


import fr.istic.gm.weplan.domain.model.dto.PageDto;
import fr.istic.gm.weplan.domain.model.dto.PageOptions;
import fr.istic.gm.weplan.domain.model.dto.RegionDto;
import fr.istic.gm.weplan.domain.service.RegionService;
import lombok.AllArgsConstructor;
import lombok.extern.slf4j.Slf4j;

@AllArgsConstructor
@Slf4j
public class RegionController {
    private RegionService regionService;

    public PageDto<RegionDto> getRegions(PageOptions pageOptions) {
        return null;
    }
}
